moved mail to service; added note about cancel account; added admin API calls for re and de activate for users
