+++ client/app/elements/page-profile/page-profile.html
@@ -25,6 +25,9 @@
       .tab-container {
         margin-top: 15px;
       }
+      .cancel {
+        font-size: 12px;
+      }
     </style>
 
     <div>
@@ -39,6 +42,9 @@
           <tr><td><span class="caps">city:</span></td> <td><span class="city caps">[[user.city]]</span></div></td></tr>
           <tr><td><span class="caps">country:</span></td> <td><span class="country">[[user.country]]</span></div></td></tr>
         </table>
+        <div class="cancel">
+          To cancel your account <a href="/contact">send us a note</a>.
+        </div>
       </div>
 
       <paper-tabs class="tab-container" selected="{{selected}}">
